THREAD.ENH: Create AsynchronousWorker thread.
